+++ src/main/java/lotto/domain/LottoStatistic.java
@@ -16,11 +16,11 @@ final class LottoStatistic {
 
     public LottoScoreDto getScore() {
         List<LottoResult> scores = calculateScore();
-        long three = getThree(scores);
-        long four = getFour(scores);
-        long five = getFive(scores);
-        long fiveBonus = getFiveBonus(scores);
-        long six = getSix(scores);
+        int three = getThree(scores);
+        int four = getFour(scores);
+        int five = getFive(scores);
+        int fiveBonus = getFiveBonus(scores);
+        int six = getSix(scores);
         double rate = getRate(three, four, five, fiveBonus, six);
         return new LottoScoreDto(three, four, five, fiveBonus, six, rate);
     }
@@ -31,44 +31,44 @@ final class LottoStatistic {
                 .collect(Collectors.toList());
     }
 
-    private long getThree(List<LottoResult> scores) {
-        return scores.stream()
+    private int getThree(List<LottoResult> scores) {
+        return (int) scores.stream()
                 .filter(it -> it == LottoResult.THREE_CORRECT)
                 .count();
     }
 
-    private long getFour(List<LottoResult> scores) {
-        return scores.stream()
+    private int getFour(List<LottoResult> scores) {
+        return (int) scores.stream()
                 .filter(it -> it == LottoResult.FOUR_CORRECT)
                 .count();
     }
 
-    private long getFive(List<LottoResult> scores) {
-        return scores.stream()
+    private int getFive(List<LottoResult> scores) {
+        return (int) scores.stream()
                 .filter(it -> it == LottoResult.FIVE_CORRECT)
                 .count();
     }
 
-    private long getFiveBonus(List<LottoResult> scores) {
-        return scores.stream()
+    private int getFiveBonus(List<LottoResult> scores) {
+        return (int) scores.stream()
                 .filter(it -> it == LottoResult.FIVE_BONUS_CORRECT)
                 .count();
     }
 
-    private long getSix(List<LottoResult> scores) {
-        return scores.stream()
+    private int getSix(List<LottoResult> scores) {
+        return (int) scores.stream()
                 .filter(it -> it == LottoResult.SIX_CORRECT)
                 .count();
     }
 
-    private double getRate(long three, long four, long five, long fiveBonus, long six) {
-        long total = three * LottoResult.THREE_CORRECT.money() +
+    private double getRate(int three, int four, int five, int fiveBonus, int six) {
+        double total = three * LottoResult.THREE_CORRECT.money() +
                 four * LottoResult.FOUR_CORRECT.money() +
                 five * LottoResult.FIVE_CORRECT.money() +
                 fiveBonus * LottoResult.FIVE_BONUS_CORRECT.money() +
                 six * LottoResult.SIX_CORRECT.money();
         int lottoPrice = LottoConstants.LOTTO_PRICE.value();
         int totalCount = lottos.size();
-        return (double) total / (lottoPrice * totalCount) * 100;
+        return total / (lottoPrice * totalCount) * 100;
     }
 }
